# api-spec/src/main/openapi/auth.yaml
openapi: 3.0.3
info:
  title: Travel Planner Authentication API
  description: API for user authentication and token management
  version: 1.0.0
  contact:
    name: Your Name
    email: your.email@example.com

servers:
  - url: http://localhost:8080
    description: Local development server

tags:
  - name: Authentication
    description: API for user authentication and token management

paths:
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Authenticate user
      description: Authenticates a user with username and password, returns JWT token and sets refresh token cookie
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequestDto'
      responses:
        '200':
          description: Successfully authenticated
          headers:
            Set-Cookie:
              description: Refresh token cookie
              schema:
                type: string
                example: refreshToken=eyJhbGc...; Path=/; HttpOnly; Max-Age=604800
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponseDto'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register new user
      description: Registers a new user with username, password, and email
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequestDto'
      responses:
        '200':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "UserEntity registered successfully"
        '400':
          description: Invalid request data or user already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Uses the refresh token cookie to generate a new access token
      operationId: refreshToken
      parameters:
        - name: Cookie
          in: header
          description: Refresh token cookie
          required: true
          schema:
            type: string
            example: "refreshToken=eyJhbGc..."
      responses:
        '200':
          description: New access token generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponseDto'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout user
      description: Invalidates the refresh token and clears the refresh token cookie
      operationId: logout
      parameters:
        - name: Cookie
          in: header
          description: Refresh token cookie
          required: true
          schema:
            type: string
            example: "refreshToken=eyJhbGc..."
      responses:
        '200':
          description: Logged out successfully
          headers:
            Set-Cookie:
              description: Cleared refresh token cookie
              schema:
                type: string
                example: refreshToken=; Path=/; HttpOnly; Max-Age=0
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Logged out successfully"
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/password/reset-request:
    post:
      tags:
        - Authentication
      summary: Request password reset
      description: Sends a password reset email to the user if the username exists
      operationId: requestPasswordReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequestDto'
      responses:
        '200':
          description: Request processed (always returns success for security)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "If the username exists, a password reset email has been sent"
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/password/reset:
    post:
      tags:
        - Authentication
      summary: Reset password
      description: Resets a user's password using a valid token
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetDto'
      responses:
        '200':
          description: Password reset successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Password has been reset successfully"
        '400':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/password/validate-token:
    get:
      tags:
        - Authentication
      summary: Validate token
      description: Checks if a password reset token is valid and not expired
      operationId: validateToken
      parameters:
        - name: token
          in: query
          description: Password reset token to validate
          required: true
          schema:
            type: string
            example: "abc123def456ghi789"
      responses:
        '200':
          description: Token validation result
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                    description: Whether the token is valid
                    example: true
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    AuthRequestDto:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          description: Username or email
          example: "john.doe@example.com"
          minLength: 3
          maxLength: 100
        password:
          type: string
          description: User password
          format: password
          example: "SecurePassword123!"
          minLength: 8
          maxLength: 100

    AuthResponseDto:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT access token
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"

    RegisterRequestDto:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
          description: Desired username
          example: "johndoe"
          minLength: 3
          maxLength: 50
          pattern: '^[a-zA-Z0-9_-]+$'
        email:
          type: string
          format: email
          description: User email address
          example: "john.doe@example.com"
          maxLength: 100
        password:
          type: string
          description: User password
          format: password
          example: "SecurePassword123!"
          minLength: 8
          maxLength: 100

    PasswordResetRequestDto:
      type: object
      required:
        - username
      properties:
        username:
          type: string
          description: Username or email for password reset
          example: "john.doe@example.com"
          minLength: 3
          maxLength: 100

    PasswordResetDto:
      type: object
      required:
        - token
        - newPassword
      properties:
        token:
          type: string
          description: Password reset token received via email
          example: "abc123def456ghi789"
        newPassword:
          type: string
          description: New password
          format: password
          example: "NewSecurePassword123!"
          minLength: 8
          maxLength: 100

    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: Timestamp of the error
          example: "2025-02-15T10:30:00Z"
        status:
          type: integer
          description: HTTP status code
          example: 400
        error:
          type: string
          description: Error type
          example: "Bad Request"
        message:
          type: string
          description: Error message
          example: "Invalid username or password"
        path:
          type: string
          description: Request path
          example: "/auth/login"